Fix audio downloading path in ActiveJob.
We cannot access Rails.root in an ActiveJob, at least not if run via Sidekiq,
because it runs in a different Docker container than the Rails app and
we have no shared directories mapped between the 2 Dockers via e.g.
volumes.
When run as in-memory Job queue, this might be different.

Furthermore, we don't need to share any directories here, because the
audio file is only needed temporarily unless attached via ActiveStorage.
Therefore, we can use the systems temp directory for temporary storage.

Signed-off-by: Daniel Schnell <dschnell@grammatek.com>